pyproject.toml
[project]
name = "fashion-segcbir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.119.1",
    "huggingface-hub>=0.35.3",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "open-clip-torch>=3.2.0",
    "openai>=2.6.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
    "uvicorn[standard]>=0.38.0",
]
